splits out websocket provider configuration into its own file
