+++ Sources/AutomergeRepo/Networking/Providers/WebSocketProvider.swift
@@ -11,29 +11,6 @@ public final class WebSocketProvider: NetworkProvider {
     /// A type that represents the configuration used to create the provider.
     public typealias ProviderConfiguration = WebSocketProviderConfiguration
 
-    /// The configuration options for a WebSocket network provider.
-    public struct WebSocketProviderConfiguration: Sendable {
-        /// A Boolean value that indicates if the provider should attempt to reconnect when it fails with an error.
-        public let reconnectOnError: Bool
-        /// The verbosity of the logs sent to the unified logging system.
-        public let logLevel: LogVerbosity
-        /// The default configuration for the WebSocket network provider.
-        ///
-        /// In the default configuration:
-        ///
-        /// - `reconnectOnError` is `true`
-        public static let `default` = WebSocketProviderConfiguration(reconnectOnError: true)
-
-        /// Creates a new WebSocket network provider configuration instance.
-        /// - Parameter reconnectOnError: A Boolean value that indicates if the provider should attempt to reconnect
-        /// when it fails with an error.
-        /// - Parameter loggingAt: The verbosity of the logs sent to the unified logging system.
-        public init(reconnectOnError: Bool, loggingAt: LogVerbosity = .errorOnly) {
-            self.reconnectOnError = reconnectOnError
-            self.logLevel = loggingAt
-        }
-    }
-
     /// The active connection for this provider.
     public var peeredConnections: [PeerConnectionInfo]
     var delegate: (any NetworkEventReceiver)?
